SeqLoggers

SeqLoggers.SeqLogger

Logging.with_logger

Logging.with_logger

SeqLoggers.flush_current_logger

SeqLoggers.flush_events

SeqLoggers.flush_global_logger

SeqLoggers.SeqLogger — Type

SeqLogger(serverUrl="http://localhost:5341", postType=Background();
               minLevel=Logging.Info, apiKey="", kwargs...)

Logger that sends log events to a Seq logging server.

Notes

The kwargs correspond to additional log event properties that can be added "globally" for a SeqLogger instance. e.g. App = "DJSON", Env = "Test" # Dev, Prod, Test, UAT, HistoryId = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"

Additional log events can also be added separately for each idividual log event @info "Event" logEventProperty="logEventValue"

Logging.with_logger — Method

Logging.with_logger(@nospecialize(f::Function), demuxLogger::TeeLogger)

Extends the method Logging.with_logger to work for a LoggingExtras.TeeLogger containing a SeqLogger.

Logging.with_logger — Method

Logging.with_logger(@nospecialize(f::Function), logger::SeqLogger)

Extends the method Logging.with_logger to work for a SeqLogger.

SeqLoggers.flush_current_logger — Method

flush_current_logger()

Post the events in the logger batch event for the logger for the current task, or the global logger if none is attached to the task.

Note

In the main moduel of Atom, the current_logger is Atom.Progress.JunoProgressLogger(). Therefore, if you set SeqLogger as a Logging.global_logger in in Atom use flush_global_logger.

SeqLoggers.flush_events — Method

flush_events(teeLogger::LoggingExtras.TeeLogger)

Extend flush_events to a work for a LoggingExtras.TeeLogger containing a SeqLogger.

SeqLoggers.flush_global_logger — Method

flush_global_logger()

Post the events in the logger batch event for the global logger.

Note

If the logger is run with Logging.with_logger, this is considered a current logger Logging.current_logger and flush_current_logger. needs to be used.