SeqLoggers

SeqLoggers.SeqLogger

Logging.with_logger

Logging.with_logger

SeqLoggers.event_property!

SeqLoggers.flush_current_logger

SeqLoggers.flush_events

SeqLoggers.flush_global_logger

SeqLoggers.SeqLogger — Type

SeqLogger(serverUrl="http://localhost:5341", postType=Serial();
          minLevel=Logging.Info,
          apiKey="",
          batchSize=10,
          kwargs...)

Logger to post log events to a Seq log server.

Inputs

serverUrl – (optional, default="http://localhost:5341") Seq server url

postType – (optional, default=Serial()) defines how HTTP posts the log events

minLevel – (optional, default=Logging.Info) minimal log level required to post events

apiKey – (optional, default="") API-key string for registered Applications

batchSize – (optional, default=10) number of log events sent to Seq server in single post

kwargs – (optional) global log event properties

Global Log Event Properties

The user can provide the logger with global log event properties by using the keyword-arguments kwargs.

SeqLogger(; App="DJSON", Env="PROD", Id="24e0d145-d385-424b-b6ec-081aa17d504a")

Local Log Event Properties

For each individual log event, additional log event properties can be added which only apply to the respective log event.

@info "Log additional user id {userId}" userId="1"

This only works, if the Logging.current_logger is of type SeqLogger.

Logging.with_logger — Method

Logging.with_logger(@nospecialize(f::Function), demuxLogger::TeeLogger)

Extends the method Logging.with_logger to work for a LoggingExtras.TeeLogger containing a SeqLogger.

Note

This constitutes as type piracy and should be treated with caution.

Logging.with_logger — Method

Logging.with_logger(@nospecialize(f::Function), logger::SeqLogger)

Extends the method Logging.with_logger to work for a SeqLogger.

SeqLoggers.event_property! — Method

event_property!(logger::SeqLogger; kwargs...)

Add one or more event properties to the list of global event properties in logger.

Example

event_property!(logger, user="Me", userId=1)

Note

If a new event property with identical name as an existing on is added with event_property!, the existing property in newEventProperties is not replaced, the new property is just added to newEventProperties. However, this still works since on the Seq side the raw post events considers the last property key as the final one if more than one has the same key.

SeqLoggers.flush_current_logger — Method

flush_current_logger()

Post the events in the logger batch event for the logger for the current task, or the global logger if none is attached to the task.

Note

In the main moduel of Atom, the current_logger is Atom.Progress.JunoProgressLogger(). Therefore, if you set SeqLogger as a Logging.global_logger in in Atom use flush_global_logger.

SeqLoggers.flush_events — Method

flush_events(teeLogger::LoggingExtras.TeeLogger)

Extend flush_events to a work for a LoggingExtras.TeeLogger containing a SeqLogger.

SeqLoggers.flush_global_logger — Method

flush_global_logger()

Post the events in the logger batch event for the global logger.

Note

If the logger is run with Logging.with_logger, this is considered a current logger Logging.current_logger and flush_current_logger. needs to be used.