SeqLoggers

SeqLoggers.SeqLogger

Logging.handle_message

Logging.with_logger

Logging.with_logger

SeqLoggers.event_property!

SeqLoggers.event_property!

SeqLoggers.flush_current_logger

SeqLoggers.flush_events

SeqLoggers.flush_events

SeqLoggers.flush_global_logger

SeqLoggers.joinurl

SeqLoggers.parse_event_str_from_args

SeqLoggers.post_log_events

SeqLoggers.stringify

SeqLoggers.SeqLogger — Method

SeqLogger(
    server_url::AbstractString;
    min_level::Logging.LogLevel=Logging.Info,
    api_key::AbstractString="",
    batch_size::Int=10,
    event_properties...
)

Logger to post log events to a Seq log server.

Inputs

server_url – Seq server url (e.g. "http://localhost:5341")

min_level – (optional, default=Logging.Info) minimal log level to filter the log events

api_key – (optional, default="") API-key string for registered Applications

batch_size – (optional, default=10) number of log events sent to Seq server in single post

event_properties – (optional) global log event properties

Global Log Event Properties

The SeqLogger constructor allows to add global log event properties to the logger using keyword-arguments.

SeqLogger("http://localhost:5341"; App="DJSON", Env="PROD", Id="24e0d145-d385-424b-b6ec-081aa17d504a")

Local Log Event Properties

For each individual log event, additional log event properties can be added which only apply to a single log event.

@info "Log additional user id {userId}" userId="1"

Note: This only works, if the Logging.current_logger is of type SeqLogger or "contains" a SeqLogger.

Logging.handle_message — Method

Logging.handle_message(logger::SeqLogger, args...; kwargs...)

Extends the function Logging.handle_message for SeqLoggers.

Note

If the event batch of the logger is "full", the log events are flushed.

A ReentrantLock is used to make the logger thread-safe.

Logging.with_logger — Method

Logging.with_logger(@nospecialize(f::Function), demuxLogger::TeeLogger)

Extends the method Logging.with_logger to work for a LoggingExtras.TeeLogger containing a SeqLogger.

Note

This constitutes as type piracy and should be treated with caution. But it is necessary such that also TeeLoggers that contain a SeqLogger do flush the log events after exiting the with_logger function.

Logging.with_logger — Method

Logging.with_logger(@nospecialize(f::Function), logger::SeqLogger)

Extends the function Logging.with_logger for SeqLoggers.

Note

After running the function f, the SeqLogger needs to flush the log events to make sure that the entire log batch is sent to the Seq server.

SeqLoggers.event_property! — Method

event_property!(teelogger::LoggingExtras.TeeLogger; kwargs...)

Extend event_property! to a work for a LoggingExtras.TeeLogger containing a SeqLogger.

SeqLoggers.event_property! — Method

event_property!(logger::SeqLogger; kwargs...)

Add one or more event properties to the list of global event properties in logger.

Example

event_property!(logger, user="Me", userId=1)

Note

If a new event property with identical name as an existing on is added with event_property!, the existing property in new_event_properties is not replaced, the new property is just added to new_event_properties. However, this still works since on the Seq side the raw post events considers the last property key as the final one if more than one has the same key.

SeqLoggers.flush_current_logger — Method

flush_current_logger()

Post the events in the logger batch event for the logger for the current task, or the global logger if none is attached to the task.

Note

In the main moduel of Atom, the current_logger is Atom.Progress.JunoProgressLogger(). Therefore, if you set SeqLogger as a Logging.global_logger in in Atom use flush_global_logger.

SeqLoggers.flush_events — Method

flush_events(teeLogger::LoggingExtras.TeeLogger)

Extend flush_events to a work for a LoggingExtras.TeeLogger containing a SeqLogger.

SeqLoggers.flush_events — Method

flush_events(logger::SeqLogger)

Post all log events contained in the event batch of logger to the Seq server and clear the event batch afterwards.

SeqLoggers.flush_global_logger — Method

flush_global_logger()

Post the events in the logger batch event for the global logger.

Note

If the logger is run with Logging.with_logger, this is considered a current logger Logging.current_logger and flush_current_logger. needs to be used.

SeqLoggers.joinurl — Method

joinurl(left::AbstractString, right::AbstractString)

Join the left and right part of a URL, by removing trailing frontslashes and add an additional frontslashes if required.

SeqLoggers.parse_event_str_from_args — Method

parse_event_str_from_args(logger::SeqLogger, message_args::NamedTuple)::String

Create a log event string from the named tuple message_args.

SeqLoggers.post_log_events — Method

post_request(logger::SeqLogger, json_str::AbstractString)

Send POST request with body json_str to Seq server.

SeqLoggers.stringify — Method

stringify(; kwargs...)

Convert keywords arguments into a string that conforms the log event message structure used in the Seq logger.